tpm2_import: use helper routines
Refactor tpm2_import to use helper routines, updating and cleansing
duplicate/incorrect structure choices along the way.
