+++ lib/tpm2_util.h
@@ -357,4 +357,5 @@ bool tpm2_util_object_save(TSS2_SYS_CONTEXT *sapi_ctx,
 bool tpm2_util_calc_unique(TPMI_ALG_HASH name_alg, TPM2B_PRIVATE_VENDOR_SPECIFIC *key,
         TPM2B_DIGEST *seed, TPM2B_DIGEST *unique);
 
+
 #endif /* STRING_BYTES_H */
